convert TestIntegration_RandomnessRequest to use simulated backend
