+++ core/internal/cltest/factories.go
@@ -171,6 +171,19 @@ func NewJobWithFluxMonitorInitiatorWithBridge() models.JobSpec {
 	return j
 }
 
+// NewJobWithRandomnessLog create new Job with VRF initiator
+func NewJobWithRandomnessLog() models.JobSpec {
+	j := NewJob()
+	j.Initiators = []models.Initiator{{
+		JobSpecID: j.ID,
+		Type:      models.InitiatorRandomnessLog,
+		InitiatorParams: models.InitiatorParams{
+			Address: NewAddress(),
+		},
+	}}
+	return j
+}
+
 // NewTx returns a Tx using a specified from address and sentAt
 func NewTx(from common.Address, sentAt uint64) *models.Tx {
 	tx := &models.Tx{
